Add Thermostat script

+++ recipes-geraldpi/thermostat/files/room_average.py
@@ -11,7 +11,7 @@ async def main():
     rooms = db.Room.unique()
     for r in rooms:
 	# Grap the last 5 reading and average them
-        print(r,df[df["Room"] == r].tail(5)["Temp"].mean(),sep=";")
+        print(r,int(df[df["Room"] == r].tail(5)["Temp"].mean()),sep=";")
 
 asyncio.run(main())
